backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
python-dotenv>=0.19.0
pydantic>=1.8.2
requests>=2.26.0
python-multipart>=0.0.5
aiosqlite>=0.17.0
httpx>=0.23.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
ratelimit>=2.2.1,<3.0.0
python-multipart>=0.0.5
bcrypt>=4.0.1
python-jose[cryptography]
